Fix regression error

+++ sdnist/metrics/regression.py
@@ -153,12 +153,12 @@ class LinearRegressionMetric:
 
         # calculate regression lines for target and synthetic data
         if self.ts.shape[0] > 1 and len(self.ts[xc].unique()) > 1:
-            self.t_reg = stats.linregress(self.ts[xc], self.ts[yc])
+            self.t_reg = stats.linregress(self.ts[xc].astype(float), self.ts[yc].astype(float))
             self.t_slope = round(self.t_reg.slope, 2)
             self.t_intercept = round(self.t_reg.intercept, 2)
 
         if self.ss.shape[0] > 1 and len(self.ss[xc].unique()) > 1:
-            self.s_reg = stats.linregress(self.ss[xc], self.ss[yc])
+            self.s_reg = stats.linregress(self.ss[xc].astype(float), self.ss[yc].astype(float))
             self.s_slope = round(self.s_reg.slope, 2)
             self.s_intercept = round(self.s_reg.intercept, 2)
